Include Sample Publisher in SampleApplication, expand readme a little.
