+++ sample_application/config/config.exs
@@ -18,7 +18,7 @@ config :kaufmann_ex,
   consumer_group: System.get_env("CONSUMER_GROUP"),
   default_topic: System.get_env("KAFKA_TOPIC"),
   producer_mod: KaufmannEx.Publisher,
-  metadata_mod: Sample.EventHandler,
+  metadata_mod: Sample.Publisher,
   schema_path: "priv/schemas",
   schema_registry_uri: System.get_env("SCHEMA_REGISTRY_PATH"),
   service_name: "SampleService",
